+++ tonic_tutorial/src/lib.rs
@@ -1,5 +1,13 @@
+//! Library root — re-exports public API and wires internal modules.
+
 pub mod data;
-pub mod proto;
+pub mod geometry;
+mod index;
 pub mod service;
 
 pub use service::RouteGuideService;
+
+// prost-generated code lives here; adjust the path to match your build.rs
+pub mod proto {
+    tonic::include_proto!("routeguide");
+}
