+++ tonic_tutorial/src/main.rs
@@ -1,21 +1,28 @@
+//! gRPC server entry-point
+//!
+//! After the library refactor, `RouteGuideService` is re-exported
+//! from `tonic_tutorial::service`, and the generated proto module
+//! is still available at `tonic_tutorial::proto`, so the old import
+//! paths continue to work.
+
 use std::net::SocketAddr;
 use tonic::transport::Server;
 use tonic_tutorial::{
-    data, // helper that deserialises route_guide_db.json
+    data, // helper: load route_guide_db.json
     proto::route_guide_server::RouteGuideServer,
-    RouteGuideService,
+    RouteGuideService, // re-exported in lib.rs
 };
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // ------------------------------------------------------------------
-    // 1 · Load the features once at start-up
+    // 1 · Load features once at start-up
     // ------------------------------------------------------------------
     let features = data::load()?; // Vec<proto::Feature>
     println!("✅ Loaded {} features", features.len());
 
     // ------------------------------------------------------------------
-    // 2 · Instantiate the service (builds maps + R-tree inside)
+    // 2 · Instantiate the service (builds DashMap + R-tree internally)
     // ------------------------------------------------------------------
     let route_guide = RouteGuideService::new(features);
 
@@ -23,7 +30,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // 3 · Launch the gRPC server
     // ------------------------------------------------------------------
     let addr: SocketAddr = "[::1]:10000".parse()?;
-    println!("🚀 Starting gRPC server on http://{}", addr);
+    println!("🚀 gRPC server listening on http://{addr}");
 
     Server::builder()
         .add_service(RouteGuideServer::new(route_guide))
