+++ tonic_tutorial/src/service/mod.rs
@@ -1,0 +1,4 @@
+//! Public entry-point for the gRPC server implementation.
+
+mod rpc;
+pub use rpc::RouteGuideService;
